Completed solution for exer# 4.21.

+++ part4/blog_list/controllers/blogs.js
@@ -15,7 +15,7 @@ blogsRouter.post('/', async (req, res, next) => {
   const body = req.body
 
   /*
-  Getting random user from users collection and using user.id to create blog.
+    Getting random user from users collection and using user.id to post a blog.
   */
   // const userArr = await User.aggregate([
   //   { $sample: { size: 1 } }
@@ -23,7 +23,7 @@ blogsRouter.post('/', async (req, res, next) => {
   // const user = await User.findById(userArr[0]._id)
 
   /*
-  Getting tokenized user.id from request authorization header.
+    Getting tokenized user.id from request authorization header.
   */
   // const token = getTokenFrom(req)
   if (!req.token) {
@@ -62,8 +62,52 @@ blogsRouter.post('/', async (req, res, next) => {
 })
 
 blogsRouter.delete('/:id', async (req, res, next) => {
+  // check if token exist in request header
+  if (!req.token){
+    return res.status(401).json({ error: 'token missing' })
+  }
+
+  let decodedToken
+  let blog
+
+  // retrieve blog and check if exist from collections
+  try {
+    blog = await Blog.findById(req.params.id)
+
+    if(!blog) {
+      return res.status(404).json({ error: 'blog does not exist' })
+    }
+  } catch(ex) {
+    next(ex)
+  }
+
+  // decode token using jsonwebtoken
+  try {
+    decodedToken = jwt.verify(req.token, process.env.SECRET)
+  } catch(ex) {
+    next(ex)
+    return res
+  }
+
+  // find and retrieve user to update the blogs array and remove the blog.id
+  const user = await User.findById(decodedToken.id)
+
+  // console.log('****blog user id:', typeof(blog.user.toString()), blog.user.toString())
+  // console.log('decoded token id:', typeof(decodedToken.id), decodedToken.id.toString())
+  // console.log('user:', user)
+  // console.log('blog:', blog)
+  // console.log('blog id', blog.id.toString())
+
+  // compare blog creator and requestor if the same user.id
+  if (blog.user.toString() !== decodedToken.id){
+    return res.status(401).json({ error: 'permission denied: user did not create blog' })
+  }
+
   try {
     await Blog.findByIdAndRemove(req.params.id)
+
+    user.blogs = user.blogs.filter(b => b.toString() !== blog.id.toString())
+    await user.save()
     res.status(204).end()
   }catch(exception) {
     next(exception)
